+++ inst/NEWS.md
@@ -3,12 +3,15 @@
  - TODO: make check Call attribute for MCA2 and postopNCK, in the combination
    it apparently loses the MCA2 call 
    (see file D15etalOptimized.RData in colbournTableReport folder)
-
+   
+04 Aug 2025, version 0.11
+ - implemented an iterative version of function coverage
+   that also allows to abort at the first non-perfect coverage
 
 31 July 2025, still version 0.10
  - documented clearly that nlevels order does not determine the output 
    column order in MCA2
- - hand over objects from the function environment to the cluster
+ - hand over objects from the function environment to the parallel cluster
    in function coverage
  - Bug fix: Ns_CK_doubling was wrong for v=3, as it did not multiply plus
    with v-1